% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/big.letter.theme.R
\name{big.letter.theme}
\alias{big.letter.theme}
\title{Create a theme with large font sizes}
\usage{
big.letter.theme(font.size)
}
\arguments{
\item{font.size}{}
}
\value{
ggplot theme
}
\description{
Create a theme with large font sizes
}

