\name{connectDB}
\alias{connectDB}
\title{Creates a connection to the DB and returns connection as a global variable, con}
\usage{
  connectDB(credentials)
}
\arguments{
  \item{credentials}{- list containing connection
  information}
}
\value{
  None - run for side-effects.
}
\description{
  Creates a connection to the DB and returns connection as
  a global variable, con
}

