% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToListEntry.R
\name{convertToListEntry}
\alias{convertToListEntry}
\title{Take a line from a genParam parameter file and turns it into a list element}
\usage{
convertToListEntry(line)
}
\arguments{
\item{line}{a line from a parameter file}
}
\value{
list selement
}
\description{
Take a line from a genParam parameter file and turns it into a list element
}
