% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HasChange.R
\name{HasChange}
\alias{HasChange}
\title{Writes an list of regression models generated via table to a file.}
\usage{
HasChange(x, k)
}
\arguments{
\item{x}{vector}

\item{k}{indicators whether there is a change}

\item{epsilon}{minimimally detectabl echange.}
}
\value{
a vector
}
\description{
Writes an list of regression models generated via table to a file.
}
