\name{JJHmisc-package}
\alias{JJHmisc-package}
\alias{JJHmisc}
\docType{package}
\title{
Collection of scripts by John Horton. 

}
\description{
Most of these are scripts to work with LaTeX tables. 

}
\details{
\tabular{ll}{
Package: \tab JJHmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-19\cr
License: \tab GPL v3\cr
}
}
\author{
John Horton

Maintainer: John Horton  <john.joseph.horton@gmail.com>

}
\references{}

\keyword{ package }
\keyword{ LaTeX }
\keyword{ Reproducible Research }
\seealso{
}
\examples{
}
