% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoteWrapper.R
\name{NoteWrapper}
\alias{NoteWrapper}
\title{Places a table note in a minitable}
\usage{
NoteWrapper(x, text.width = 0.9)
}
\arguments{
\item{text.width}{the with of the page to use up (0 to 1.0)}
}
\description{
Places a table note in a minitable
}

