% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.big.number.R
\name{pp.big.number}
\alias{pp.big.number}
\title{Formats a large number with commas.}
\usage{
pp.big.number(e)
}
\arguments{
\item{e}{the number.}
}
\value{
the number with comments.
}
\description{
Formats a large number with commas.
}
