% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadParamFile.R
\name{ReadParamFile}
\alias{ReadParamFile}
\title{Reads in a parameter file constructed from a genParam file}
\usage{
ReadParamFile(file.name)
}
\arguments{
\item{file.name}{the file where the parameters}
}
\value{
a list with keys and values corresponding to the parameter file
}
\description{
Reads in a parameter file constructed from a genParam file
}
