\name{regression.table}
\alias{regression.table}
\title{Writes an list of regression models generated via table to a file.}
\usage{
regression.table(models, renames, out.file)
}
\arguments{
  \item{models}{is a list of models.}

  \item{renames}{is a list of renames.}

  \item{out.file}{path to where table shoul be written}
}
\value{
Nothing - writes to file out.file
}
\description{
Writes an list of regression models generated via table to
a file.
}

