% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render}
\alias{render}
\title{Creata line for a LaTeX codebook.}
\usage{
render(x, lst)
}
\arguments{
\item{x}{tempalte}

\item{lst}{a hash table list of replacement rules}
}
\value{
filled-in template
}
\description{
Creata line for a LaTeX codebook.
}

