\name{spanning.latex.column.label}
\alias{spanning.latex.column.label}
\title{Creates a spanning multicolumn for a LaTeX table.}
\usage{
  spanning.latex.column.label(num.columns, width, label)
}
\arguments{
  \item{num.columns}{is the number of columns the
  multicolumn should span}

  \item{width}{is the width in cm}

  \item{label}{is the label for the column.}
}
\value{
  text snippet.
}
\description{
  Creates a spanning multicolumn for a LaTeX table.
}

