% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starLabel.R
\name{starLabel}
\alias{starLabel}
\title{Returns significance stars from a p-value}
\usage{
starLabel(p)
}
\arguments{
\item{p}{p-value}
}
\value{
symbol
}
\description{
Returns significance stars from a p-value
}
