% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableContinuous2.R
\name{tableContinuous2}
\alias{tableContinuous2}
\title{Modified version of the reporttools code}
\usage{
tableContinuous2(vars, weights = NA, subset = NA, group = NA,
  stats = c("n", "min", "q1", "median", "mean", "q3", "max", "s", "iqr",
  "na"), prec = 1, col.tit = NA, col.tit.font = c("bf", "", "sf", "it",
  "rm"), print.pval = c("none", "anova", "kruskal"), pval.bound = 10^-4,
  declare.zero = 10^-10, cap = "", lab = "", font.size = "footnotesize",
  longtable = TRUE, disp.cols = NA, nams = NA)
}
\arguments{
\item{vars}{}
}
\value{
line to be written.
}
\description{
Modified version of the reporttools code
}

