% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableNominal2.R
\name{tableNominal2}
\alias{tableNominal2}
\title{Modified version of the codetools function}
\usage{
tableNominal2(vars, weights = NA, subset = NA, group = NA,
  miss.cat = NA, print.pval = c("none", "fisher", "chi2"),
  pval.bound = 10^-4, fisher.B = 2000, vertical = TRUE, cap = "",
  lab = "", col.tit.font = c("bf", "", "sf", "it", "rm"),
  font.size = "footnotesize", longtable = TRUE, nams = NA,
  cumsum = TRUE)
}
\arguments{
\item{vars}{}
}
\value{
line to be written.
}
\description{
Modified version of the codetools function
}

