\name{wasItUsed}
\alias{wasItUsed}
\title{For each variable in a dataset, tests whether or not that variable was used.}
\usage{
wasItUsed(r.file, data.frame)
}
\arguments{
  \item{f.file}{}

  \item{data.frame}{}

  \item{note}{is a note.}
}
\value{
line to be written.
}
\description{
For each variable in a dataset, tests whether or not that
variable was used.
}

