\name{write.image}
\alias{write.image}
\title{Writes a plot as an image to some specificed location.}
\usage{
write.image(filename, g, width = 5, height = 5, format = "png")
}
\arguments{
  \item{filename}{where to write the file}

  \item{g}{is the plot}

  \item{width}{width}

  \item{heigh}{height}

  \item{format}{is the file format to use.}
}
\value{
None - run for side-effects.
}
\description{
Writes a plot as an image to some specificed location.
}

