% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEntry.R
\name{writeEntry}
\alias{writeEntry}
\title{Writes a key value pair to a LaTeX parameter file}
\usage{
writeEntry(key, value, parameter.file)
}
\arguments{
\item{key}{the name of the LaTeX variable}

\item{value}{the value of the variable}

\item{parameter.file}{the paramater file}
}
\description{
Writes a key value pair to a LaTeX parameter file
}
