% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeImage.R
\name{writeImage}
\alias{writeImage}
\title{Write images and a tex file for input into a document}
\usage{
writeImage(g, file.name.no.ext, width = 15, height = 8,
  path = "../../writeup/plots/", position = "h", line.width = "0.8",
  caption = "", notes = "", include.tex.wrapper = FALSE)
}
\arguments{
\item{g}{graphic}

\item{file.name.no.ext}{file name without an extension}

\item{width}{in inches}

\item{height}{in inches}

\item{path}{path to direction where image should be written}

\item{position}{figure position for tex}

\item{line.width}{fraction of line width for tex}

\item{caption}{the caption for tex}

\item{notes}{the notes for tex}
}
\value{
None
}
\description{
Write images and a tex file for input into a document
}
