% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteListToParamFile.R
\name{WriteListToParamFile}
\alias{WriteListToParamFile}
\title{Writes a list of all parameters to a file}
\usage{
WriteListToParamFile(full.list, file.name)
}
\value{
none
}
\description{
Writes a list of all parameters to a file
}
