\name{JJHmisc-package}
\alias{JJHmisc-package}
\alias{JJHmisc}
\docType{package}
\title{
What the package does (short line)

}
\description{
More about what it does (maybe more than one line)

}
\details{
\tabular{ll}{
Package: \tab JJHmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-19\cr
License: \tab What license is it under?\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{

}

\keyword{ package }
\seealso{
}
\examples{
}
