% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasItUsed.R
\name{wasItUsed}
\alias{wasItUsed}
\title{For each variable in a dataset, tests whether or not that variable was used.}
\usage{
wasItUsed(r.file, data.frame)
}
\arguments{
\item{note}{is a note.}
}
\value{
line to be written.
}
\description{
For each variable in a dataset, tests whether or not that variable was used.
}
