% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.image.R
\name{write.image}
\alias{write.image}
\title{Writes a plot as an image to some specificed location.}
\usage{
write.image(filename, g, width = 5, height = 5, format = "png")
}
\arguments{
\item{filename}{where to write the file}

\item{g}{is the plot}

\item{width}{width}

\item{format}{is the file format to use.}

\item{heigh}{height}
}
\value{
None - run for side-effects.
}
\description{
Writes a plot as an image to some specificed location.
}

