\name{writeImage}
\alias{writeImage}
\title{Write images}
\usage{
writeImage(g, file.name.no.ext, width = 15, height = 8,
  path = "../../writeup/plots/")
}
\arguments{
  \item{g}{graphic}

  \item{file.name.no.ext}{file name without an extension}

  \item{width}{in inches}

  \item{height}{in inches}

  \item{path}{path to direction where image should be
  written}
}
\value{
None
}
\description{
Write images
}

